refactor: add Global Exception Handler with used @RestControllerAdvice
